+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-09-23 02:50:11
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 74c1b924e5b380e8c485a5069deaca51146ef94a
+
+## [ForwardWidgets-2kuai] - 2025-09-23 02:50:11
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 6b06cf5c0ef1658ab0f9eed28524862f7441aede
+
+## [ForwardWidgets-pack1r] - 2025-09-23 02:50:11
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: d242db63f883ffc49445d7d0316402a5565c8bb2
+
+
 ## [ForwardWidgets-huangxd] - 2025-09-22 02:59:59
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main